requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
joblib
jupyter
streamlit
torch
autogluon.tabular
lightgbm
xgboost
catboost
fastai
